Update packages to avoid split-package problem.
