+++ codegen-plugin/src/main/kotlin/io/spine/examples/protodata/hello/plugin/BuilderBeforeReturnRenderer.kt
@@ -23,7 +23,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.examples.protodata.hello
+package io.spine.examples.protodata.hello.plugin
 
 import io.spine.protodata.renderer.Renderer
 import io.spine.protodata.renderer.SourceFile
