+++ codegen-plugin/src/main/kotlin/io/spine/examples/protodata/hello/plugin/BuilderValidationMethods.kt
@@ -23,7 +23,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.examples.protodata.hello
+package io.spine.examples.protodata.hello.plugin
 
 import com.google.common.collect.ArrayListMultimap
 import java.nio.file.Path
