+++ codegen-plugin/src/main/kotlin/io/spine/examples/protodata/hello/plugin/SizeOptionViewRepository.kt
@@ -23,45 +23,31 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.examples.protodata.hello
+package io.spine.examples.protodata.hello.plugin
 
-import io.spine.protodata.plugin.Plugin
+import io.spine.core.EventContext
+import io.spine.protodata.event.FieldOptionDiscovered
 import io.spine.protodata.plugin.ViewRepository
-import io.spine.protodata.renderer.Renderer
+import io.spine.server.route.EventRoute
+import io.spine.server.route.EventRouting
 
 /**
- * ProtoData [Plugin] that renders validation code for the [ArrayOfSizeOption]
- * which is applied to a repeated field.
- *
- * Value of the option is an expression that may refer to other fields
- * of the message type and supports basic math operations,
- * such as `+`, `-`, `*`, `/`.
- *
- * Example of definition:
- * ```
- * message Board {
- *     repeated Cell cell = 1 [(required) = true,
- *         (size).value = "side_size * side_size"];
- *
- *     int32 side_size = 2 [(required) = true, (min).value = "3"];
- * }
- * ```
+ * The repository for [SizeOptionView].
  */
-public class ApplySizeOptionPlugin : Plugin {
-
-    override fun renderers(): List<Renderer<*>> {
-        // A container for collecting generated validation methods, which should
-        // later be added to the `build` method of the message builder class.
-        val builderValidationMethods = BuilderValidationMethods()
-
-        return listOf(
-            ValidateSizeOptionRenderer(builderValidationMethods),
-            BuilderBeforeReturnPrinter(),
-            BuilderBeforeReturnRenderer(builderValidationMethods)
-        )
-    }
-
-    override fun viewRepositories(): Set<ViewRepository<*, *, *>> {
-        return setOf(SizeOptionViewRepository())
+internal class SizeOptionViewRepository : ViewRepository<SizeOptionId,
+        SizeOptionView,
+        SizeOption>() {
+    override fun setupEventRouting(routing: EventRouting<SizeOptionId>) {
+        super.setupEventRouting(routing)
+        routing.route(FieldOptionDiscovered::class.java)
+        { message: FieldOptionDiscovered, _: EventContext? ->
+            EventRoute.withId(
+                sizeOptionId {
+                    filePath = message.file
+                    typeName = message.type
+                    fieldName = message.field
+                }
+            )
+        }
     }
 }
