--- codegen-plugin/src/main/proto/spine/examples/protodata/hello/size_option.proto
@@ -1,60 +1,0 @@
-/*
- * Copyright 2024, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-syntax = "proto3";
-
-package spine.examples.protodata.hello;
-
-import "spine/options.proto";
-
-option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.examples.protodata.hello";
-option java_outer_classname = "SizeOptionProto";
-option java_multiple_files = true;
-
-import "spine/protodata/ast.proto";
-
-// View on the `ArrayOfSizeOption`.
-message SizeOption {
-  option (entity).kind = PROJECTION;
-
-  // ID of the option.
-  SizeOptionId id = 1;
-
-  // Value of the option which is a validation expression
-  // for the size of a repeated field.
-  string expression = 2;
-}
-
-// Triplet File+Type+Field unambiguously identifies the option.
-message SizeOptionId {
-
-  spine.protodata.FilePath file_path = 1;
-
-  spine.protodata.TypeName type_name = 2;
-
-  spine.protodata.FieldName field_name = 3;
-}
