+++ integration-tests/build.gradle.kts
@@ -47,7 +47,7 @@ apply {
 protoData {
     // Run ProtoData with the `size` option plugin enabled.
     plugins(
-        "io.spine.examples.protodata.hello.ApplySizeOptionPlugin"
+        "io.spine.examples.protodata.hello.plugin.ApplySizeOptionPlugin"
     )
 }
