+++ integration-tests/src/test/resources/negative-cases/empty_expression_value.proto
@@ -26,16 +26,16 @@
 
 syntax = "proto3";
 
-package spine.examples.protodata.hello;
+package spine.examples.protodata.hello.test;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.examples.protodata.hello";
+option java_package = "io.spine.examples.protodata.hello.test";
 option java_outer_classname = "SizeOptionTestMsgProto";
 option java_multiple_files = true;
 
-import "spine/examples/protodata/hello/options.proto";
+import "spine/examples/protodata/hello/option/options.proto";
 
 message SizeOptionTestMsg {
